Feature/vcf and maxdepth support (#27)
* Add VCF support and max-depth support

This adds support for taking a BCF/VCF file to specify the positions of
interest. It also allows a user to specify the max depth in order to
help control runtime in cases where there is very high coverage.

This also includes a bugfix to `only-depth` that was skipping records
that started on the stop of a region, or stopped on the start of a
region.

Lastly, this adds the full contig name to the logs for better debuging,
and clarifies when tid is output instead of the contig name.